had to change parseInt to parseFloat to allow decimals for the updated scores

+++ frontend/src/routes/Leaderboard.jsx
@@ -71,7 +71,7 @@ const Leaderboard = () => {
     setUsers((preUsers) => {
       return preUsers.map((user) => {
         if (user.id === userId) {
-          return { ...user, score: parseInt(event.target.value) };
+          return { ...user, score: parseFloat(event.target.value) };
         } else {
           return user;
         }
